+++ src/com/projectkorra/projectkorra/ability/FireAbility.java
@@ -20,7 +20,7 @@ import com.projectkorra.projectkorra.Element;
 import com.projectkorra.projectkorra.GeneralMethods;
 import com.projectkorra.projectkorra.ProjectKorra;
 import com.projectkorra.projectkorra.Element.SubElement;
-import com.projectkorra.projectkorra.ability.lighting.LightEmitTask;
+import com.projectkorra.projectkorra.ability.lighting.FauxLight;
 import com.projectkorra.projectkorra.ability.util.Collision;
 import com.projectkorra.projectkorra.configuration.ConfigManager;
 import com.projectkorra.projectkorra.util.ParticleEffect;
@@ -160,10 +160,10 @@ public abstract class FireAbility extends ElementalAbility {
 	public void playFirebendingParticles(final Location loc, final int amount, final double xOffset, final double yOffset, final double zOffset) {
 		if (this.getBendingPlayer().canUseSubElement(SubElement.BLUE_FIRE)) {
 			ParticleEffect.SOUL_FIRE_FLAME.display(loc, amount, xOffset, yOffset, zOffset);
-			new LightEmitTask(loc.getBlock(), 15, 16L); // Light for blue flames.
+			new FauxLight(loc.getBlock(), 15, 16L); // Light for blue flames.
 		} else {
 			ParticleEffect.FLAME.display(loc, amount, xOffset, yOffset, zOffset);
-			new LightEmitTask(loc.getBlock(), 13, 16L); // Light for regular flames.
+			new FauxLight(loc.getBlock(), 13, 16L); // Light for regular flames.
 		}
 	}
 
@@ -189,7 +189,7 @@ public abstract class FireAbility extends ElementalAbility {
 
 	public static void playLightningbendingParticle(final Location loc, final double xOffset, final double yOffset, final double zOffset) {
 		GeneralMethods.displayColoredParticle("#01E1FF", loc, 1, xOffset, yOffset, zOffset);
-		new LightEmitTask(loc.getBlock(), 15, 16L); // Light for lightning abilities.
+		new FauxLight(loc.getBlock(), 15, 16L); // Light for lightning abilities.
 	}
 
 	public static void playLightningbendingSound(final Location loc) {
